+++ src/entities/Scoreboard.js
