+++ src/entities/Stars.js
